* Update models and forms to be compatible with Django 1.11
